@@ -12,17 +12,20 @@
     "@icons-pack/react-simple-icons": "^9.3.0",
     "@react-three/drei": "^9.115.0",
     "@react-three/fiber": "^8.17.10",
+    "@react-three/postprocessing": "^2.16.3",
+    "@react-three/rapier": "^1.5.0",
     "@types/three": "^0.169.0",
     "axios": "^1.7.7",
     "clsx": "^2.1.1",
-    "next": "^14.1.0",
+    "ecctrl": "^1.0.90",
+    "next": "^14.2.16",
     "next-themes": "^0.2.1",
     "react": "^18.2.0",
     "react-dom": "18.2.0",
     "react-rough-notation": "^1.0.5",
     "tailwind-merge": "^2.5.4",
     "three": "^0.169.0",
-    "wouter": "^3.3.5"
+    "uuid-by-string": "^4.0.0"
   },
   "devDependencies": {
     "autoprefixer": "^10.4.12",
