@@ -1,259 +1,206 @@
-import React, { useEffect, useState } from "react";
+"use client";
+import { useEffect, useState } from "react";
 import Link from "next/link";
 import { useTheme } from "next-themes";
-import { useRouter } from "next/router";
+import { usePathname } from "next/navigation";
 import userData from "@constants/data";
 
 export default function Navbar() {
-    const router = useRouter();
-    const { theme, setTheme } = useTheme();
-    const [mounted, setMounted] = useState(false);
+  const { theme, setTheme } = useTheme();
+  const [mounted, setMounted] = useState(false);
 
-    useEffect(() => {
-        setMounted(true);
-    }, []);
+  useEffect(() => {
+    setMounted(true);
+  }, []);
 
-    return (
-        <nav className="max-w-6xl mx-auto px-4 py-10 ">
-            <div className="flex  md:flex-row justify-between items-center">
-                <div className="flex flex-col">
-                    <Link href="/">
-                        <h1 className="font-semibold text-xl dark:text-gray-100">
-                            {userData.name}
-                        </h1>
-                        <p className="text-base font-light text-gray-500 dark:text-gray-300">
-                            {userData.designation}
-                        </p>
-                    </Link>
-                </div>
+  const navigation = [
+    { name: "About", href: "/about" },
+    { name: "Projects", href: "/projects" },
+    { name: "Experience", href: "/experience" },
+    { name: "Contact", href: "/contact" },
+  ];
 
-                <ul className="space-x-8 hidden md:flex  ">
-                    <li>
-                        <Link
-                            href="/about"
-                            className={`text-base  ${
-                                router.asPath === "/about"
-                                    ? "text-gray-800 font-bold dark:text-gray-400"
-                                    : "text-gray-600 dark:text-gray-300 font-normal "
-                            }`}
-                        >
-                            About{" "}
-                            {router.asPath === "/about" && (
-                                <svg
-                                    xmlns="http://www.w3.org/2000/svg"
-                                    width="16"
-                                    height="16"
-                                    fill="currentColor"
-                                    className="bi bi-arrow-down inline-block h-3 w-3"
-                                    viewBox="0 0 16 16"
-                                >
-                                    <path
-                                        fillRule="evenodd"
-                                        d="M8 1a.5.5 0 0 1 .5.5v11.793l3.146-3.147a.5.5 0 0 1 .708.708l-4 4a.5.5 0 0 1-.708 0l-4-4a.5.5 0 0 1 .708-.708L7.5 13.293V1.5A.5.5 0 0 1 8 1z"
-                                    />
-                                </svg>
-                            )}
-                        </Link>
-                    </li>
-                    <li>
-                        <Link
-                            href="/projects"
-                            className={`text-base  ${
-                                router.asPath === "/projects"
-                                    ? "text-gray-800 font-bold dark:text-gray-400"
-                                    : "text-gray-600 dark:text-gray-300 font-normal "
-                            }`}
-                        >
-                            Projects
-                            {router.asPath === "/projects" && (
-                                <svg
-                                    xmlns="http://www.w3.org/2000/svg"
-                                    width="16"
-                                    height="16"
-                                    fill="currentColor"
-                                    className="bi bi-arrow-down inline-block h-3 w-3"
-                                    viewBox="0 0 16 16"
-                                >
-                                    <path
-                                        fillRule="evenodd"
-                                        d="M8 1a.5.5 0 0 1 .5.5v11.793l3.146-3.147a.5.5 0 0 1 .708.708l-4 4a.5.5 0 0 1-.708 0l-4-4a.5.5 0 0 1 .708-.708L7.5 13.293V1.5A.5.5 0 0 1 8 1z"
-                                    />
-                                </svg>
-                            )}
-                        </Link>
-                    </li>
-                    <li>
-                        <Link
-                            href="/experience"
-                            className={`text-base  ${
-                                router.asPath === "/experience"
-                                    ? "text-gray-800 font-bold dark:text-gray-400"
-                                    : "text-gray-600 dark:text-gray-300 font-normal "
-                            }`}
-                        >
-                            Experience{" "}
-                            {router.asPath === "/experience" && (
-                                <svg
-                                    xmlns="http://www.w3.org/2000/svg"
-                                    width="16"
-                                    height="16"
-                                    fill="currentColor"
-                                    className="bi bi-arrow-down inline-block h-3 w-3"
-                                    viewBox="0 0 16 16"
-                                >
-                                    <path
-                                        fillRule="evenodd"
-                                        d="M8 1a.5.5 0 0 1 .5.5v11.793l3.146-3.147a.5.5 0 0 1 .708.708l-4 4a.5.5 0 0 1-.708 0l-4-4a.5.5 0 0 1 .708-.708L7.5 13.293V1.5A.5.5 0 0 1 8 1z"
-                                    />
-                                </svg>
-                            )}
-                        </Link>
-                    </li>
-                    <li>
-                        <Link
-                            href="/contact"
-                            className={`text-base  ${
-                                router.asPath === "/contact"
-                                    ? "text-gray-800 font-bold dark:text-gray-400"
-                                    : "text-gray-600 dark:text-gray-300 font-normal "
-                            }`}
-                        >
-                            Contact
-                            {router.asPath === "/contact" && (
-                                <svg
-                                    xmlns="http://www.w3.org/2000/svg"
-                                    width="16"
-                                    height="16"
-                                    fill="currentColor"
-                                    className="bi bi-arrow-down inline-block h-3 w-3"
-                                    viewBox="0 0 16 16"
-                                >
-                                    <path
-                                        fillRule="evenodd"
-                                        d="M8 1a.5.5 0 0 1 .5.5v11.793l3.146-3.147a.5.5 0 0 1 .708.708l-4 4a.5.5 0 0 1-.708 0l-4-4a.5.5 0 0 1 .708-.708L7.5 13.293V1.5A.5.5 0 0 1 8 1z"
-                                    />
-                                </svg>
-                            )}
-                        </Link>
-                    </li>
-                </ul>
+  const socialIcons = {
+    github: (
+      <svg
+        xmlns="http://www.w3.org/2000/svg"
+        width="24"
+        height="24"
+        fill="currentColor"
+        className="h-5 w-5"
+        viewBox="0 0 24 24"
+        aria-hidden="true"
+      >
+        <path d="M12 .297c-6.63 0-12 5.373-12 12 0 5.303 3.438 9.8 8.205 11.385.6.113.82-.258.82-.577 0-.285-.01-1.04-.015-2.04-3.338.724-4.042-1.61-4.042-1.61C4.422 18.07 3.633 17.7 3.633 17.7c-1.087-.744.084-.729.084-.729 1.205.084 1.838 1.236 1.838 1.236 1.07 1.835 2.809 1.305 3.495.998.108-.776.417-1.305.76-1.605-2.665-.3-5.466-1.332-5.466-5.93 0-1.31.465-2.38 1.235-3.22-.135-.303-.54-1.523.105-3.176 0 0 1.005-.322 3.3 1.23.96-.267 1.98-.399 3-.405 1.02.006 2.04.138 3 .405 2.28-1.552 3.285-1.23 3.285-1.23.645 1.653.24 2.873.12 3.176.765.84 1.23 1.91 1.23 3.22 0 4.61-2.805 5.625-5.475 5.92.42.36.81 1.096.81 2.22 0 1.606-.015 2.896-.015 3.286 0 .315.21.69.825.57C20.565 22.092 24 17.592 24 12.297c0-6.627-5.373-12-12-12" />
+      </svg>
+    ),
+    instagram: (
+      <svg
+        xmlns="http://www.w3.org/2000/svg"
+        width="16"
+        height="16"
+        fill="currentColor"
+        className="h-5 w-5"
+        viewBox="0 0 16 16"
+        aria-hidden="true"
+      >
+        <path d="M8 0C5.829 0 5.556.01 4.703.048 3.85.088 3.269.222 2.76.42a3.917 3.917 0 0 0-1.417.923A3.927 3.927 0 0 0 .42 2.76C.222 3.268.087 3.85.048 4.7.01 5.555 0 5.827 0 8.001c0 2.172.01 2.444.048 3.297.04.852.174 1.433.372 1.942.205.526.478.972.923 1.417.444.445.89.719 1.416.923.51.198 1.09.333 1.942.372C5.555 15.99 5.827 16 8 16s2.444-.01 3.298-.048c.851-.04 1.434-.174 1.943-.372a3.916 3.916 0 0 0 1.416-.923c.445-.445.718-.891.923-1.417.197-.509.332-1.09.372-1.942C15.99 10.445 16 10.173 16 8s-.01-2.445-.048-3.299c-.04-.851-.175-1.433-.372-1.941a3.926 3.926 0 0 0-.923-1.417A3.911 3.911 0 0 0 13.24.42c-.51-.198-1.092-.333-1.943-.372C10.443.01 10.172 0 7.998 0h.003zm-.717 1.442h.718c2.136 0 2.389.007 3.232.046.78.035 1.204.166 1.486.275.373.145.64.319.92.599.28.28.453.546.598.92.11.281.24.705.275 1.485.039.843.047 1.096.047 3.231s-.008 2.389-.047 3.232c-.035.78-.166 1.203-.275 1.485a2.47 2.47 0 0 1-.599.919c-.28.28-.546.453-.92.598-.28.11-.704.24-1.485.276-.843.038-1.096.047-3.232.047s-2.39-.009-3.233-.047c-.78-.036-1.203-.166-1.485-.276a2.478 2.478 0 0 1-.92-.598 2.48 2.48 0 0 1-.6-.92c-.109-.281-.24-.705-.275-1.485-.038-.843-.046-1.096-.046-3.233 0-2.136.008-2.388.046-3.231.036-.78.166-1.204.276-1.486.145-.373.319-.64.599-.92.28-.28.546-.453.92-.598.282-.11.705-.24 1.485-.276.738-.034 1.024-.044 2.515-.045v.002zm4.988 1.328a.96.96 0 1 0 0 1.92.96.96 0 0 0 0-1.92zm-4.27 1.122a4.109 4.109 0 1 0 0 8.217 4.109 4.109 0 0 0 0-8.217zm0 1.441a2.667 2.667 0 1 1 0 5.334 2.667 2.667 0 0 1 0-5.334z" />
+      </svg>
+    ),
+    linkedin: (
+      <svg
+        xmlns="http://www.w3.org/2000/svg"
+        width="16"
+        height="16"
+        fill="currentColor"
+        className="h-5 w-5"
+        viewBox="0 0 16 16"
+        aria-hidden="true"
+      >
+        <path d="M0 1.146C0 .513.526 0 1.175 0h13.65C15.474 0 16 .513 16 1.146v13.708c0 .633-.526 1.146-1.175 1.146H1.175C.526 16 0 15.487 0 14.854V1.146zm4.943 12.248V6.169H2.542v7.225h2.401zm-1.2-8.212c.837 0 1.358-.554 1.358-1.248-.015-.709-.52-1.248-1.342-1.248-.822 0-1.359.54-1.359 1.248 0 .694.521 1.248 1.327 1.248h.016zm4.908 8.212V9.359c0-.216.016-.432.08-.586.173-.431.568-.878 1.232-.878.869 0 1.216.662 1.216 1.634v3.865h2.401V9.25c0-2.22-1.184-3.252-2.764-3.252-1.274 0-1.845.7-2.165 1.193v.025h-.016a5.54 5.54 0 0 1 .016-.025V6.169h-2.4c.03.678 0 7.225 0 7.225h2.4z" />
+      </svg>
+    ),
+  };
 
-                <div className="space-x-4 flex flex-row items-center">
-                    <Link
-                        href={userData.socialLinks.github}
-                        className="text-base font-normal text-gray-600 dark:text-gray-300"
-                        target="_blank"
-                        aria-label="GitHub"
-                        role="link"
-                    >
-                        <svg
-                            xmlns="http://www.w3.org/2000/svg"
-                            width="24"
-                            height="24"
-                            fill="currentColor"
-                            className="bi bi-twitter h-5 w-5"
-                            viewBox="0 0 24 24"
-                        >
-                            <path d="M12 .297c-6.63 0-12 5.373-12 12 0 5.303 3.438 9.8 8.205 11.385.6.113.82-.258.82-.577 0-.285-.01-1.04-.015-2.04-3.338.724-4.042-1.61-4.042-1.61C4.422 18.07 3.633 17.7 3.633 17.7c-1.087-.744.084-.729.084-.729 1.205.084 1.838 1.236 1.838 1.236 1.07 1.835 2.809 1.305 3.495.998.108-.776.417-1.305.76-1.605-2.665-.3-5.466-1.332-5.466-5.93 0-1.31.465-2.38 1.235-3.22-.135-.303-.54-1.523.105-3.176 0 0 1.005-.322 3.3 1.23.96-.267 1.98-.399 3-.405 1.02.006 2.04.138 3 .405 2.28-1.552 3.285-1.23 3.285-1.23.645 1.653.24 2.873.12 3.176.765.84 1.23 1.91 1.23 3.22 0 4.61-2.805 5.625-5.475 5.92.42.36.81 1.096.81 2.22 0 1.606-.015 2.896-.015 3.286 0 .315.21.69.825.57C20.565 22.092 24 17.592 24 12.297c0-6.627-5.373-12-12-12" />
-                        </svg>
-                    </Link>
-                    <Link
-                        href={userData.socialLinks.instagram}
-                        className="text-base font-normal text-gray-600 dark:text-gray-300"
-                        target="_blank"
-                        aria-label="Instagram"
-                        role="link"
-                    >
-                        <svg
-                            xmlns="http://www.w3.org/2000/svg"
-                            width="16"
-                            height="16"
-                            fill="currentColor"
-                            className="bi bi-instagram h-5 w-5"
-                            viewBox="0 0 16 16"
-                        >
-                            <path d="M8 0C5.829 0 5.556.01 4.703.048 3.85.088 3.269.222 2.76.42a3.917 3.917 0 0 0-1.417.923A3.927 3.927 0 0 0 .42 2.76C.222 3.268.087 3.85.048 4.7.01 5.555 0 5.827 0 8.001c0 2.172.01 2.444.048 3.297.04.852.174 1.433.372 1.942.205.526.478.972.923 1.417.444.445.89.719 1.416.923.51.198 1.09.333 1.942.372C5.555 15.99 5.827 16 8 16s2.444-.01 3.298-.048c.851-.04 1.434-.174 1.943-.372a3.916 3.916 0 0 0 1.416-.923c.445-.445.718-.891.923-1.417.197-.509.332-1.09.372-1.942C15.99 10.445 16 10.173 16 8s-.01-2.445-.048-3.299c-.04-.851-.175-1.433-.372-1.941a3.926 3.926 0 0 0-.923-1.417A3.911 3.911 0 0 0 13.24.42c-.51-.198-1.092-.333-1.943-.372C10.443.01 10.172 0 7.998 0h.003zm-.717 1.442h.718c2.136 0 2.389.007 3.232.046.78.035 1.204.166 1.486.275.373.145.64.319.92.599.28.28.453.546.598.92.11.281.24.705.275 1.485.039.843.047 1.096.047 3.231s-.008 2.389-.047 3.232c-.035.78-.166 1.203-.275 1.485a2.47 2.47 0 0 1-.599.919c-.28.28-.546.453-.92.598-.28.11-.704.24-1.485.276-.843.038-1.096.047-3.232.047s-2.39-.009-3.233-.047c-.78-.036-1.203-.166-1.485-.276a2.478 2.478 0 0 1-.92-.598 2.48 2.48 0 0 1-.6-.92c-.109-.281-.24-.705-.275-1.485-.038-.843-.046-1.096-.046-3.233 0-2.136.008-2.388.046-3.231.036-.78.166-1.204.276-1.486.145-.373.319-.64.599-.92.28-.28.546-.453.92-.598.282-.11.705-.24 1.485-.276.738-.034 1.024-.044 2.515-.045v.002zm4.988 1.328a.96.96 0 1 0 0 1.92.96.96 0 0 0 0-1.92zm-4.27 1.122a4.109 4.109 0 1 0 0 8.217 4.109 4.109 0 0 0 0-8.217zm0 1.441a2.667 2.667 0 1 1 0 5.334 2.667 2.667 0 0 1 0-5.334z" />
-                        </svg>
-                    </Link>
+  return (
+    <header>
+      <nav className="max-w-6xl mx-auto px-4 py-10" aria-label="Main navigation">
+        <div className="flex md:flex-row justify-between items-center">
+          {/* Logo/Name Section */}
+          <div className="flex flex-col">
+            <Link href="/" className="hover:opacity-80 transition-opacity">
+              <h1 className="font-semibold text-xl dark:text-gray-100">{userData.name}</h1>
+              <p className="text-base font-light text-gray-500 dark:text-gray-300">
+                {userData.designation}
+              </p>
+            </Link>
+          </div>
 
-                    <Link
-                        href={userData.socialLinks.linkedin}
-                        className="text-base font-normal text-gray-600 dark:text-gray-300"
-                        target="_blank"
-                        aria-label="LinkedIn"
-                        role="link"
-                    >
-                        <svg
-                            xmlns="http://www.w3.org/2000/svg"
-                            width="16"
-                            height="16"
-                            fill="currentColor"
-                            className="bi bi-linkedin h-5 w-5"
-                            viewBox="0 0 16 16"
-                        >
-                            <path d="M0 1.146C0 .513.526 0 1.175 0h13.65C15.474 0 16 .513 16 1.146v13.708c0 .633-.526 1.146-1.175 1.146H1.175C.526 16 0 15.487 0 14.854V1.146zm4.943 12.248V6.169H2.542v7.225h2.401zm-1.2-8.212c.837 0 1.358-.554 1.358-1.248-.015-.709-.52-1.248-1.342-1.248-.822 0-1.359.54-1.359 1.248 0 .694.521 1.248 1.327 1.248h.016zm4.908 8.212V9.359c0-.216.016-.432.08-.586.173-.431.568-.878 1.232-.878.869 0 1.216.662 1.216 1.634v3.865h2.401V9.25c0-2.22-1.184-3.252-2.764-3.252-1.274 0-1.845.7-2.165 1.193v.025h-.016a5.54 5.54 0 0 1 .016-.025V6.169h-2.4c.03.678 0 7.225 0 7.225h2.4z" />
-                        </svg>
-                    </Link>
-                    <button
-                        aria-label="Toggle Dark Mode"
-                        type="button"
-                        className="w-10 h-10 p-3 rounded focus:outline-none"
-                        onClick={() => setTheme(theme === "dark" ? "light" : "dark")}
-                    >
-                        {mounted && (
-                            <svg
-                                xmlns="http://www.w3.org/2000/svg"
-                                viewBox="0 0 24 24"
-                                fill="currentColor"
-                                stroke="currentColor"
-                                className="w-4 h-4 text-yellow-500 dark:text-yellow-500"
-                            >
-                                {theme === "dark" ? (
-                                    <path
-                                        strokeLinecap="round"
-                                        strokeLinejoin="round"
-                                        strokeWidth={2}
-                                        d="M12 3v1m0 16v1m9-9h-1M4 12H3m15.364 6.364l-.707-.707M6.343 6.343l-.707-.707m12.728 0l-.707.707M6.343 17.657l-.707.707M16 12a4 4 0 11-8 0 4 4 0 018 0z"
-                                    />
-                                ) : (
-                                    <path
-                                        strokeLinecap="round"
-                                        strokeLinejoin="round"
-                                        strokeWidth={2}
-                                        d="M20.354 15.354A9 9 0 018.646 3.646 9.003 9.003 0 0012 21a9.003 9.003 0 008.354-5.646z"
-                                    />
-                                )}
-                            </svg>
-                        )}
-                    </button>
-                </div>
-            </div>
-            <div className="space-x-8 block md:hidden mt-4">
-                <Link
-                    href="/about"
-                    className="text-base font-normal text-gray-600 dark:text-gray-300"
+          {/* Desktop Navigation */}
+          <ul className="space-x-8 hidden md:flex">
+            {navigation.map((item) => (
+              <li key={item.href}>
+                <NavLink href={item.href}>{item.name}</NavLink>
+              </li>
+            ))}
+          </ul>
+
+          {/* Social Links & Theme Toggle */}
+          <div className="space-x-4 flex flex-row items-center">
+            <SocialLink
+              href={userData.socialLinks.github}
+              label="GitHub Profile"
+              icon={socialIcons.github}
+            />
+            <SocialLink
+              href={userData.socialLinks.instagram}
+              label="Instagram Profile"
+              icon={socialIcons.instagram}
+            />
+            <SocialLink
+              href={userData.socialLinks.linkedin}
+              label="LinkedIn Profile"
+              icon={socialIcons.linkedin}
+            />
+
+            {/* Theme Toggle Button */}
+            <button
+              aria-label={`Switch themes`}
+              type="button"
+              className="w-10 h-10 p-3 rounded  transition-transform hover:scale-110"
+              onClick={() => setTheme(theme === "dark" ? "light" : "dark")}
+            >
+              {mounted && (
+                <svg
+                  xmlns="http://www.w3.org/2000/svg"
+                  viewBox="0 0 24 24"
+                  fill="currentColor"
+                  stroke="currentColor"
+                  className="w-4 h-4 text-yellow-500 dark:text-yellow-500"
+                  aria-hidden="true"
                 >
-                    About
-                </Link>
-                <Link
-                    href="/projects"
-                    className="text-base font-normal text-gray-600 dark:text-gray-300"
-                >
-                    Projects
-                </Link>
-                <Link
-                    href="/experience"
-                    className="text-base font-normal text-gray-600 dark:text-gray-300"
-                >
-                    Experience
-                </Link>
-                <Link
-                    href="/contact"
-                    className="text-base font-normal text-gray-600 dark:text-gray-300"
-                >
-                    Contact
-                </Link>
-            </div>
-        </nav>
-    );
+                  {theme === "dark" ? (
+                    <path
+                      strokeLinecap="round"
+                      strokeLinejoin="round"
+                      strokeWidth={2}
+                      d="M12 3v1m0 16v1m9-9h-1M4 12H3m15.364 6.364l-.707-.707M6.343 6.343l-.707-.707m12.728 0l-.707.707M6.343 17.657l-.707.707M16 12a4 4 0 11-8 0 4 4 0 018 0z"
+                    />
+                  ) : (
+                    <path
+                      strokeLinecap="round"
+                      strokeLinejoin="round"
+                      strokeWidth={2}
+                      d="M20.354 15.354A9 9 0 018.646 3.646 9.003 9.003 0 0012 21a9.003 9.003 0 008.354-5.646z"
+                    />
+                  )}
+                </svg>
+              )}
+            </button>
+          </div>
+        </div>
+
+        {/* Mobile Navigation */}
+        <div className="space-x-8 block md:hidden mt-4">
+          {navigation.map((item) => (
+            <Link
+              key={item.href}
+              href={item.href}
+              className="text-base font-normal text-gray-600 dark:text-gray-300 hover:text-gray-800 dark:hover:text-gray-200 transition-colors"
+            >
+              {item.name}
+            </Link>
+          ))}
+        </div>
+      </nav>
+    </header>
+  );
+}
+
+function NavLink({ href, children }) {
+  const pathname = usePathname();
+  const isActive = pathname === href;
+
+  return (
+    <Link
+      href={href}
+      className={`text-base transition-colors ${
+        isActive
+          ? "text-gray-800 font-bold dark:text-gray-400"
+          : "text-gray-600 dark:text-gray-300 font-normal hover:text-gray-800 dark:hover:text-gray-200"
+      }`}
+    >
+      {children}
+      {isActive && (
+        <svg
+          xmlns="http://www.w3.org/2000/svg"
+          width="16"
+          height="16"
+          fill="currentColor"
+          className="bi bi-arrow-down inline-block h-3 w-3 ml-1"
+          viewBox="0 0 16 16"
+          aria-hidden="true"
+        >
+          <path
+            fillRule="evenodd"
+            d="M8 1a.5.5 0 0 1 .5.5v11.793l3.146-3.147a.5.5 0 0 1 .708.708l-4 4a.5.5 0 0 1-.708 0l-4-4a.5.5 0 0 1 .708-.708L7.5 13.293V1.5A.5.5 0 0 1 8 1z"
+          />
+        </svg>
+      )}
+    </Link>
+  );
+}
+
+function SocialLink({ href, label, icon: Icon }) {
+  return (
+    <Link
+      href={href}
+      className="text-base font-normal text-gray-600 dark:text-gray-300 hover:text-gray-800 dark:hover:text-gray-200 transition-colors"
+      target="_blank"
+      rel="noopener noreferrer"
+      aria-label={label}
+    >
+      {Icon}
+    </Link>
+  );
 }
