@@ -1,369 +1,374 @@
 const userData = {
-    githubUsername: "kabaskill",
-    name: "Oguz Kabasakal",
-    designation: "Developer | Designer",
-    avatarUrl: "/avatar.png",
-    email: "oguzkabasakal@gmail.com",
-    phone: "+49 1523 678 6051",
-    address: "Darmstadt, Germany",
-    projects: {
-        dev: [
-            {
-                title: "Fugue State - Coming Soon",
-                link: "https://github.com/kabaskill/fugue-state-react",
-                imgUrl: "/FugueState-1.png",
-            },
-            {
-                title: "GameDalf",
-                link: "https://gamedalf.vercel.app",
-                imgUrl: "/gamedalf.png",
-            },
-            {
-                title: "Tourio",
-                link: "https://tourio-travel-app-rho.vercel.app",
-                imgUrl: "/tourio.png",
-            },
-            {
-                title: "Rubby the Duck - TBD",
-                link: "https://www.youtube.com/watch?v=tXc_zhsvte8",
-                imgUrl: "/tourio.png",
-            },
-            {
-                title: "Unity Demo - Bedroom Guitarist",
-                link: "https://play.unity.com/en/games/fb8b88f1-7bd2-4f53-a42c-da6188ae5296/bedroom-guitarist-clicker",
-                imgUrl: "/made-with-unity.jpg",
-            },
-            {
-                title: "Unity Demo - Angry Birds Clone",
-                link: "https://play.unity.com/en/games/36fb7d2c-9c59-4df0-82fa-2faaa65e145e/mad-birds",
-                imgUrl: "/made-with-unity.jpg",
-            },
-            {
-                title: "p5.js Demo - Snake Clone",
-                link: "https://editor.p5js.org/kabaskill/full/-Ax07LXQt",
-                imgUrl: "/p5js-1.png",
-            },
-            {
-                title: "p5.js Demo - Game of Life",
-                link: "https://editor.p5js.org/kabaskill/full/U_uirRDt3",
-                imgUrl: "/p5js-4.png",
-            },
-            {
-                title: "p5.js Demo - Clock v2",
-                link: "https://editor.p5js.org/kabaskill/full/j7CA0UMDu",
-                imgUrl: "/p5js-2.png",
-            },
-            {
-                title: "p5.js Demo - Clock v1",
-                link: "https://editor.p5js.org/kabaskill/full/cOf__xkYk",
-                imgUrl: "/p5js-3.png",
-            },
-        ],
-
-        sound: [
-            {
-                title: "Showreel",
-                link: "https://www.youtube.com/watch?v=WdwXRmtowxA",
-                imgUrl: "",
-            },
-            {
-                title: "Metanoia",
-                link: "https://www.youtube.com/watch?v=Qif1lN3WjEo",
-                imgUrl: "",
-            },
-            {
-                title: "SAYA",
-                link: "https://www.youtube.com/watch?v=U7g0K1s2__w",
-                imgUrl: "",
-            },
-            {
-                title: "Tapari-Spielwelt",
-                link: "https://www.youtube.com/watch?v=VWLYl_1PPBI",
-                imgUrl: "",
-            },
-            {
-                title: "Logo Design",
-                link: "https://www.youtube.com/watch?v=ILklZxQckEM",
-                imgUrl: "",
-            },
-            {
-                title: "CON-B Podcast",
-                link: "https://open.spotify.com/show/2GrAkzQqxH0tOgU0ZlI1bk?go=1&sp_cid=145cc92d75f1249fc8cfdcf8522f6cb1&utm_source=embed_player_p&utm_medium=desktop&nd=1&dlsi=54e6c5f455514c02",
-                imgUrl: "/Podcast.jpg",
-            },
-            {
-                title: "An Ordinary Day",
-                link: "https://www.youtube.com/watch?v=ttb41edWA7U",
-                imgUrl: "",
-            },
-            {
-                title: "Soul of the Light",
-                link: "https://vimeo.com/242382390",
-                imgUrl: "",
-            },
-            {
-                title: "Gamble",
-                link: "https://vimeo.com/104676877",
-                imgUrl: "",
-            },
-            {
-                title: "Artern - Lost Generation",
-                link: "https://www.youtube.com/watch?v=_KtUbHRzklU",
-                imgUrl: "",
-            },
-            {
-                title: "Caravan: Travels of an Ancient Civilization",
-                link: "https://www.youtube.com/watch?v=YtIUmmLDSOE",
-                imgUrl: "",
-            },
-            {
-                title: "Benzersiz",
-                link: "https://www.youtube.com/watch?v=kIDfSPAmMKU",
-                imgUrl: "",
-            },
-            {
-                title: "Mevlana's 744th Birthday",
-                link: "https://www.youtube.com/watch?v=pZcudV-PUYY",
-                imgUrl: "",
-            },
-            {
-                title: "Philips Infomercials",
-                link: "https://youtu.be/WdwXRmtowxA?t=301",
-                imgUrl: "/philips.png",
-            },
-            {
-                title: "Aksu Candy",
-                link: "https://www.youtube.com/watch?v=UsbmHNWdSf0",
-                imgUrl: "",
-            },
-            {
-                title: "Human: Unknown",
-                link: "https://www.youtube.com/watch?v=Qmc0SBNCqVA",
-                imgUrl: "",
-            },
-            {
-                title: "Zero",
-                link: "https://vimeo.com/287928387",
-                imgUrl: "",
-            },
-            {
-                title: "Redesign: KIA",
-                link: "https://vimeo.com/100425716",
-                imgUrl: "",
-            },
-            {
-                title: "Redesign: Subaru",
-                link: "https://vimeo.com/100425715",
-                imgUrl: "",
-            },
-            {
-                title: "Redesign: UP!",
-                link: "https://vimeo.com/100407058",
-                imgUrl: "",
-            },
-            {
-                title: "10th Village",
-                link: "https://www.youtube.com/watch?v=epgUqTY49Is",
-                imgUrl: "",
-            },
-            {
-                title: "Klein Verhaal Over Een Muzikant",
-                link: "https://vimeo.com/100407055",
-                imgUrl: "",
-            },
-            {
-                title: "Odemis Yem Infomercial",
-                link: "https://www.youtube.com/watch?v=TM2zVTFPnis",
-                imgUrl: "",
-            },
-        ],
-
-        music: [
-            {
-                title: "Metanoia",
-                link: "https://www.youtube.com/watch?v=9lTTIfjbPy8",
-                imgUrl: "",
-            },
-            {
-                title: "SAYA",
-                link: "https://www.youtube.com/watch?v=Z_-ZmBUz1lk",
-                imgUrl: "",
-            },
-            {
-                title: "Blöf - Deli Gibi Sevdim",
-                link: "https://www.youtube.com/watch?v=aFwXboxvJ3U",
-                imgUrl: "",
-            },
-            {
-                title: "Blöf - Huysuz",
-                link: "https://www.youtube.com/watch?v=0-9jZ_arQzg",
-                imgUrl: "",
-            },
-            {
-                title: "Blöf - Olmaz",
-                link: "https://www.youtube.com/watch?v=a64I2vC2shM",
-                imgUrl: "",
-            },
-            {
-                title: "Blöf - Saklambaç",
-                link: "https://www.youtube.com/watch?v=38mLrlco9sc",
-                imgUrl: "",
-            },
-            {
-                title: "Arpedduo - Clubbed To Death",
-                link: "https://www.youtube.com/watch?v=2-PlFeWuRho",
-                imgUrl: "",
-            },
-            {
-                title: "Arpedduo - Metallica Medley",
-                link: "https://www.youtube.com/watch?v=1MldLlHLVNs",
-                imgUrl: "",
-            },
-            {
-                title: "Arpedduo - Godfather Theme",
-                link: "https://www.youtube.com/watch?v=6nZNDKQsbTM",
-                imgUrl: "",
-            },
-            {
-                title: "Arpedduo - Wild Flower",
-                link: "https://www.youtube.com/watch?v=l7pab1ZCs30",
-                imgUrl: "",
-            },
-            {
-                title: "Arpedduo - Zülüf",
-                link: "https://www.youtube.com/watch?v=dLfDpRs09j0",
-                imgUrl: "",
-            },
-            {
-                title: "Arpedduo - Çiçek Abbas",
-                link: "https://www.youtube.com/watch?v=je4oMZtpeoA",
-                imgUrl: "",
-            },
-            {
-                title: "Arpedduo - Dilek Taşı",
-                link: "https://www.youtube.com/watch?v=4hJjh-5jDaA",
-                imgUrl: "",
-            },
-            {
-                title: "Arpedduo - Neşeli Günler",
-                link: "https://www.youtube.com/watch?v=rPg7ww1XgDc",
-                imgUrl: "",
-            },
-            {
-                title: "Personal Projects - Interstellar",
-                link: "https://www.youtube.com/watch?v=3dH9OMpFQyo",
-                imgUrl: "",
-            },
-            {
-                title: "Personal Projects - It's a Step",
-                link: "https://www.youtube.com/watch?v=gfMw66pOE0o",
-                imgUrl: "",
-            },
-            {
-                title: "Personal Projects - Voyager",
-                link: "https://www.youtube.com/watch?v=cokWGEE-_nU",
-                imgUrl: "",
-            },
-            {
-                title: "Personal Projects - Serenity",
-                link: "https://www.youtube.com/watch?v=_RsxazXU0r0",
-                imgUrl: "",
-            },
-            {
-                title: "Personal Projects - Don't Shut Up",
-                link: "https://www.youtube.com/watch?v=k-gAELzwNTI",
-                imgUrl: "",
-            },
-            {
-                title: "Personal Projects - Decoy",
-                link: "https://www.youtube.com/watch?v=iy63vrUqCwg",
-                imgUrl: "",
-            },
-            {
-                title: "Personal Projects - Paradox",
-                link: "https://www.youtube.com/watch?v=Q-T8Jaz0jpA",
-                imgUrl: "",
-            },
-            {
-                title: "Personal Projects - It's Getting Hot",
-                link: "https://www.youtube.com/watch?v=uqTzyIHl37c",
-                imgUrl: "",
-            },
-            {
-                title: "Personal Projects - Gimme Shelter",
-                link: "https://www.youtube.com/watch?v=Cv43Gospk8k",
-                imgUrl: "",
-            },
-            {
-                title: "Personal Projects - Cliffs of Dover",
-                link: "https://www.youtube.com/watch?v=GryhaHChXyg",
-                imgUrl: "",
-            },
-        ],
-    },
-
-    about: {
-        title: "Developing by day, slaying demons by night",
-        description: [
-            `I've been designing soundscapes and music for all kinds of visual content since 2012.`,
-            `Studying for a Master's Degree on Expanded Media at Hochschule Darmstadt.`,
-            `Nowadays, I focused my efforts to software and game development. Started learning Unity and C#in 2019 but web development became prevalent and now I'm building full-stack web apps using various technologies.`,
-        ],
-        currentProject: "Fugue State",
-        currentProjectUrl: "https://github.com/kabaskill/fugue-state-react",
-    },
-
-    experience: [
-        {
-            title: "Web Designer / Social Media Manager",
-            company: "Formenwerkstatt GmbH",
-            year: "2024 - Present",
-            companyLink: "https://www.formenwerksttat.de/",
-            desc: "Running and managing the company's digital assets and social media, including the web site, social media accounts and internal file server.",
-        },
-        {
-            title: "Trainee - Web Developer",
-            company: "neuefische GmbH",
-            year: "2023 - 2024",
-            companyLink: "https://www.neuefische.de/",
-            desc: "540 Hours of Full-time Intensive Coding Bootcamp on Full Stack Web Development",
-        },
-        {
-            title: "Sound Designer",
-            company: "freelance",
-            year: "2011 - Present",
-            companyLink: "",
-            desc: "Designing soundscapes for all kinds of visual media and providing solutions to live recording, foley, voice-over, sound design and everything audio related in general",
-        },
-        {
-            title: "Music Teacher",
-            company: "Ministry of Education of Turkey",
-            year: "2019 - 2022",
-            companyLink: "https://www.meb.gov.tr/",
-            desc: "Teaching music theory and guitar to middle and high school students in 24 student classrooms as well as private lessons",
-        },
-        {
-            title: "Sound Supervisor",
-            company: "Erka Media",
-            year: "2017- 2019",
-            companyLink: "https://erkamedya.com/",
-            desc: "Teaching music theory and guitar to middle and high school students in 24 student classrooms as well as private lessons",
-        },
-        {
-            title: "Localization",
-            company: "Philips Turkey | Avantgarde Media",
-            year: "2014- 2017",
-            companyLink: "",
-            desc: "Teaching music theory and guitar to middle and high school students in 24 student classrooms as well as private lessons",
-        },
+  githubUsername: "kabaskill",
+  name: "Oguz Kabasakal",
+  designation: "Developer | Designer",
+  avatarUrl: "/avatar.png",
+  email: "oguzkabasakal@gmail.com",
+  phone: "+49 1523 678 6051",
+  address: "Darmstadt, Germany",
+  projects: {
+    dev: [
+      {
+        title: "Formenwerkstatt Website",
+        link: "https://frormenwerkstatt.de",
+        imgUrl: "/fw-banner.png",
+      },
+      {
+        title: "Fugue State - Coming Soon",
+        link: "https://github.com/kabaskill/fugue-state-react",
+        imgUrl: "/FugueState-1.png",
+      },
+      {
+        title: "GameDalf",
+        link: "https://gamedalf.vercel.app",
+        imgUrl: "/gamedalf.png",
+      },
+      {
+        title: "Tourio",
+        link: "https://tourio-travel-app-rho.vercel.app",
+        imgUrl: "/tourio.png",
+      },
+      {
+        title: "Rubby the Duck - TBD",
+        link: "https://www.youtube.com/watch?v=tXc_zhsvte8",
+        imgUrl: "/rubby-the-duck.png",
+      },
+      {
+        title: "Unity Demo - Bedroom Guitarist",
+        link: "https://play.unity.com/en/games/fb8b88f1-7bd2-4f53-a42c-da6188ae5296/bedroom-guitarist-clicker",
+        imgUrl: "/made-with-unity.jpg",
+      },
+      {
+        title: "Unity Demo - Angry Birds Clone",
+        link: "https://play.unity.com/en/games/36fb7d2c-9c59-4df0-82fa-2faaa65e145e/mad-birds",
+        imgUrl: "/made-with-unity.jpg",
+      },
+      {
+        title: "p5.js Demo - Snake Clone",
+        link: "https://editor.p5js.org/kabaskill/full/-Ax07LXQt",
+        imgUrl: "/p5js-1.png",
+      },
+      {
+        title: "p5.js Demo - Game of Life",
+        link: "https://editor.p5js.org/kabaskill/full/U_uirRDt3",
+        imgUrl: "/p5js-4.png",
+      },
+      {
+        title: "p5.js Demo - Clock v2",
+        link: "https://editor.p5js.org/kabaskill/full/j7CA0UMDu",
+        imgUrl: "/p5js-2.png",
+      },
+      {
+        title: "p5.js Demo - Clock v1",
+        link: "https://editor.p5js.org/kabaskill/full/cOf__xkYk",
+        imgUrl: "/p5js-3.png",
+      },
     ],
-    
-    resumeUrl: "https://drive.google.com/file/d/1HK_yHw09zytdS7vFxSy1lgLoj5Vn7nQm/view?usp=sharing",
-    socialLinks: {
-        instagram: "https://instagram.com/mr.kabaskill",
-        linkedin: "https://linkedin.com/in/oguzkabasakal",
-        github: "https://github.com/kabaskill",
+
+    sound: [
+      {
+        title: "Showreel",
+        link: "https://www.youtube.com/watch?v=WdwXRmtowxA",
+        imgUrl: "",
+      },
+      {
+        title: "Metanoia",
+        link: "https://www.youtube.com/watch?v=Qif1lN3WjEo",
+        imgUrl: "",
+      },
+      {
+        title: "SAYA",
+        link: "https://www.youtube.com/watch?v=U7g0K1s2__w",
+        imgUrl: "",
+      },
+      {
+        title: "Tapari-Spielwelt",
+        link: "https://www.youtube.com/watch?v=VWLYl_1PPBI",
+        imgUrl: "",
+      },
+      {
+        title: "Logo Design",
+        link: "https://www.youtube.com/watch?v=ILklZxQckEM",
+        imgUrl: "",
+      },
+      {
+        title: "CON-B Podcast",
+        link: "https://open.spotify.com/show/2GrAkzQqxH0tOgU0ZlI1bk?go=1&sp_cid=145cc92d75f1249fc8cfdcf8522f6cb1&utm_source=embed_player_p&utm_medium=desktop&nd=1&dlsi=54e6c5f455514c02",
+        imgUrl: "/Podcast.jpg",
+      },
+      {
+        title: "An Ordinary Day",
+        link: "https://www.youtube.com/watch?v=ttb41edWA7U",
+        imgUrl: "",
+      },
+      {
+        title: "Soul of the Light",
+        link: "https://vimeo.com/242382390",
+        imgUrl: "",
+      },
+      {
+        title: "Gamble",
+        link: "https://vimeo.com/104676877",
+        imgUrl: "",
+      },
+      {
+        title: "Artern - Lost Generation",
+        link: "https://www.youtube.com/watch?v=_KtUbHRzklU",
+        imgUrl: "",
+      },
+      {
+        title: "Caravan: Travels of an Ancient Civilization",
+        link: "https://www.youtube.com/watch?v=YtIUmmLDSOE",
+        imgUrl: "",
+      },
+      {
+        title: "Benzersiz",
+        link: "https://www.youtube.com/watch?v=kIDfSPAmMKU",
+        imgUrl: "",
+      },
+      {
+        title: "Mevlana's 744th Birthday",
+        link: "https://www.youtube.com/watch?v=pZcudV-PUYY",
+        imgUrl: "",
+      },
+      {
+        title: "Philips Infomercials",
+        link: "https://youtu.be/WdwXRmtowxA?t=301",
+        imgUrl: "/philips.png",
+      },
+      {
+        title: "Aksu Candy",
+        link: "https://www.youtube.com/watch?v=UsbmHNWdSf0",
+        imgUrl: "",
+      },
+      {
+        title: "Human: Unknown",
+        link: "https://www.youtube.com/watch?v=Qmc0SBNCqVA",
+        imgUrl: "",
+      },
+      {
+        title: "Zero",
+        link: "https://vimeo.com/287928387",
+        imgUrl: "",
+      },
+      {
+        title: "Redesign: KIA",
+        link: "https://vimeo.com/100425716",
+        imgUrl: "",
+      },
+      {
+        title: "Redesign: Subaru",
+        link: "https://vimeo.com/100425715",
+        imgUrl: "",
+      },
+      {
+        title: "Redesign: UP!",
+        link: "https://vimeo.com/100407058",
+        imgUrl: "",
+      },
+      {
+        title: "10th Village",
+        link: "https://www.youtube.com/watch?v=epgUqTY49Is",
+        imgUrl: "",
+      },
+      {
+        title: "Klein Verhaal Over Een Muzikant",
+        link: "https://vimeo.com/100407055",
+        imgUrl: "",
+      },
+      {
+        title: "Odemis Yem Infomercial",
+        link: "https://www.youtube.com/watch?v=TM2zVTFPnis",
+        imgUrl: "",
+      },
+    ],
+
+    music: [
+      {
+        title: "Metanoia",
+        link: "https://www.youtube.com/watch?v=9lTTIfjbPy8",
+        imgUrl: "",
+      },
+      {
+        title: "SAYA",
+        link: "https://www.youtube.com/watch?v=Z_-ZmBUz1lk",
+        imgUrl: "",
+      },
+      {
+        title: "Blöf - Deli Gibi Sevdim",
+        link: "https://www.youtube.com/watch?v=aFwXboxvJ3U",
+        imgUrl: "",
+      },
+      {
+        title: "Blöf - Huysuz",
+        link: "https://www.youtube.com/watch?v=0-9jZ_arQzg",
+        imgUrl: "",
+      },
+      {
+        title: "Blöf - Olmaz",
+        link: "https://www.youtube.com/watch?v=a64I2vC2shM",
+        imgUrl: "",
+      },
+      {
+        title: "Blöf - Saklambaç",
+        link: "https://www.youtube.com/watch?v=38mLrlco9sc",
+        imgUrl: "",
+      },
+      {
+        title: "Arpedduo - Clubbed To Death",
+        link: "https://www.youtube.com/watch?v=2-PlFeWuRho",
+        imgUrl: "",
+      },
+      {
+        title: "Arpedduo - Metallica Medley",
+        link: "https://www.youtube.com/watch?v=1MldLlHLVNs",
+        imgUrl: "",
+      },
+      {
+        title: "Arpedduo - Godfather Theme",
+        link: "https://www.youtube.com/watch?v=6nZNDKQsbTM",
+        imgUrl: "",
+      },
+      {
+        title: "Arpedduo - Wild Flower",
+        link: "https://www.youtube.com/watch?v=l7pab1ZCs30",
+        imgUrl: "",
+      },
+      {
+        title: "Arpedduo - Zülüf",
+        link: "https://www.youtube.com/watch?v=dLfDpRs09j0",
+        imgUrl: "",
+      },
+      {
+        title: "Arpedduo - Çiçek Abbas",
+        link: "https://www.youtube.com/watch?v=je4oMZtpeoA",
+        imgUrl: "",
+      },
+      {
+        title: "Arpedduo - Dilek Taşı",
+        link: "https://www.youtube.com/watch?v=4hJjh-5jDaA",
+        imgUrl: "",
+      },
+      {
+        title: "Arpedduo - Neşeli Günler",
+        link: "https://www.youtube.com/watch?v=rPg7ww1XgDc",
+        imgUrl: "",
+      },
+      {
+        title: "Personal Projects - Interstellar",
+        link: "https://www.youtube.com/watch?v=3dH9OMpFQyo",
+        imgUrl: "",
+      },
+      {
+        title: "Personal Projects - It's a Step",
+        link: "https://www.youtube.com/watch?v=gfMw66pOE0o",
+        imgUrl: "",
+      },
+      {
+        title: "Personal Projects - Voyager",
+        link: "https://www.youtube.com/watch?v=cokWGEE-_nU",
+        imgUrl: "",
+      },
+      {
+        title: "Personal Projects - Serenity",
+        link: "https://www.youtube.com/watch?v=_RsxazXU0r0",
+        imgUrl: "",
+      },
+      {
+        title: "Personal Projects - Don't Shut Up",
+        link: "https://www.youtube.com/watch?v=k-gAELzwNTI",
+        imgUrl: "",
+      },
+      {
+        title: "Personal Projects - Decoy",
+        link: "https://www.youtube.com/watch?v=iy63vrUqCwg",
+        imgUrl: "",
+      },
+      {
+        title: "Personal Projects - Paradox",
+        link: "https://www.youtube.com/watch?v=Q-T8Jaz0jpA",
+        imgUrl: "",
+      },
+      {
+        title: "Personal Projects - It's Getting Hot",
+        link: "https://www.youtube.com/watch?v=uqTzyIHl37c",
+        imgUrl: "",
+      },
+      {
+        title: "Personal Projects - Gimme Shelter",
+        link: "https://www.youtube.com/watch?v=Cv43Gospk8k",
+        imgUrl: "",
+      },
+      {
+        title: "Personal Projects - Cliffs of Dover",
+        link: "https://www.youtube.com/watch?v=GryhaHChXyg",
+        imgUrl: "",
+      },
+    ],
+  },
+
+  about: {
+    title: "Developing by day, slaying demons by night",
+    description: [
+      `I've been designing soundscapes and music for all kinds of visual content since 2012.`,
+      `Studying for a Master's Degree on Expanded Media at Hochschule Darmstadt.`,
+      `Nowadays, I focused my efforts to software and game development. Started learning Unity and C#in 2019 but web development became prevalent and now I'm building full-stack web apps using various technologies.`,
+    ],
+    currentProject: "Fugue State",
+    currentProjectUrl: "https://github.com/kabaskill/fugue-state-react",
+  },
+
+  experience: [
+    {
+      title: "Web Designer / Social Media Manager",
+      company: "Formenwerkstatt GmbH",
+      year: "2024 - Present",
+      companyLink: "https://www.formenwerksttat.de/",
+      desc: "Running and managing the company's digital assets and social media, including the web site, social media accounts and internal file server.",
     },
+    {
+      title: "Trainee - Web Developer",
+      company: "neuefische GmbH",
+      year: "2023 - 2024",
+      companyLink: "https://www.neuefische.de/",
+      desc: "540 Hours of Full-time Intensive Coding Bootcamp on Full Stack Web Development",
+    },
+    {
+      title: "Sound Designer",
+      company: "freelance",
+      year: "2011 - Present",
+      companyLink: "",
+      desc: "Designing soundscapes for all kinds of visual media and providing solutions to live recording, foley, voice-over, sound design and everything audio related in general",
+    },
+    {
+      title: "Music Teacher",
+      company: "Ministry of Education of Turkey",
+      year: "2019 - 2022",
+      companyLink: "https://www.meb.gov.tr/",
+      desc: "Teaching music theory and guitar to middle and high school students in 24 student classrooms as well as private lessons",
+    },
+    {
+      title: "Sound Supervisor",
+      company: "Erka Media",
+      year: "2017- 2019",
+      companyLink: "https://erkamedya.com/",
+      desc: "Teaching music theory and guitar to middle and high school students in 24 student classrooms as well as private lessons",
+    },
+    {
+      title: "Localization",
+      company: "Philips Turkey | Avantgarde Media",
+      year: "2014- 2017",
+      companyLink: "",
+      desc: "Teaching music theory and guitar to middle and high school students in 24 student classrooms as well as private lessons",
+    },
+  ],
+
+  resumeUrl: "https://drive.google.com/file/d/1HK_yHw09zytdS7vFxSy1lgLoj5Vn7nQm/view?usp=sharing",
+  socialLinks: {
+    instagram: "https://instagram.com/mr.kabaskill",
+    linkedin: "https://linkedin.com/in/oguzkabasakal",
+    github: "https://github.com/kabaskill",
+  },
 };
 
 export default userData;
