@@ -1,9 +1,9 @@
 import axios from "axios";
 
-const getLatestRepos = async (data, token) => {
+export const getLatestRepos = async (data, token) => {
   try {
     const username = data.githubUsername;
-
+    
     if (token) {
       const res = await axios.get(
         `https://api.github.com/user/repos?affiliation=owner,collaborator,organization_member&sort=pushed&per_page=6`,
@@ -13,17 +13,17 @@ const getLatestRepos = async (data, token) => {
           },
         }
       );
-      let latestSixRepos = res.data;
-      return latestSixRepos;
-    } else {
-      const res = await axios.get(
-        `https://api.github.com/users/${username}/repos?affiliation=owner,collaborator,organization_member&sort=pushed&per_page=6`
-      );
-      let latestSixRepos = res.data;
-      return latestSixRepos;
-    }
+      return res.data;
+    } 
+    
+    const res = await axios.get(
+      `https://api.github.com/users/${username}/repos?affiliation=owner,collaborator,organization_member&sort=pushed&per_page=6`
+    );
+    return res.data;
+    
   } catch (err) {
-    console.log(err);
+    console.error("Error fetching GitHub repos:", err);
+    return [];  // Return empty array in case of error
   }
 };
 
